requirements.txt
redis==4.4.2
elasticsearch[async]==8.10.1
orjson==3.10.2
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
pytest==7.4.3
requests==2.31.0
fastapi==0.111.0
pydantic==2.7.1
uvicorn==0.29.0
pydantic_settings==2.2.1
gunicorn==22.0.0
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
python-dotenv==1.0.1
starlette==0.37.2
aiohttp==3.8.6